backend/requirements.txt
fastapi[all]==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
httpx==0.27.0
groq==0.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
gunicorn==21.2.0
